+++ Plugin/CustomPosters/src/Patches/StartOfRoundPatch.cs
@@ -1,0 +1,15 @@
+using HarmonyLib;
+
+namespace CustomPosters.Patches
+{
+    [HarmonyPatch(typeof(StartOfRound))]
+    internal class StartOfRoundPatch
+    {
+        [HarmonyPostfix]
+        [HarmonyPatch("Start")]
+        private static void OnRoundStartPatch(StartOfRound __instance)
+        {
+            PosterManager.OnRoundStart(__instance);
+        }
+    }
+}